+++ apps/designer-standalone/src/app/Designer/designer.tsx
@@ -12,6 +12,7 @@ import {
   BaseFunctionService,
   BaseAppServiceService,
 } from '@microsoft/designer-client-services-logic-apps';
+import type { ContentType } from '@microsoft/designer-client-services-logic-apps';
 import { DesignerProvider, BJSWorkflowProvider, Designer } from '@microsoft/logic-apps-designer';
 import { ResourceIdentityType } from '@microsoft/utils-logic-apps';
 import { useEffect } from 'react';
@@ -97,6 +98,8 @@ const appServiceService = new BaseAppServiceService({
 
 const workflowService = { getCallbackUrl: () => Promise.resolve({ method: 'POST', value: 'Dummy url' }) };
 
+const hostService = { fetchAndDisplayContent: (title: string, url: string, type: ContentType) => console.log(title, url, type) };
+
 export const DesignerWrapper = () => {
   const { workflowDefinition, readOnly, monitoringView, darkMode, consumption, connections, runInstance } = useSelector(
     (state: RootState) => state.workflowLoader
@@ -111,6 +114,7 @@ export const DesignerWrapper = () => {
       functionService,
       appServiceService,
       workflowService,
+      hostService,
     },
     readOnly,
     isMonitoringView: monitoringView,