+++ libs/data-mapper/src/lib/ui/DataMapperDesigner.tsx
@@ -20,6 +20,7 @@ import { redoDataMapOperation, saveDataMap, showNotification, undoDataMapOperati
 import type { AppDispatch, RootState } from '../core/state/Store';
 import { convertToMapDefinition } from '../mapDefinitions';
 import { LogCategory, LogService } from '../utils/Logging.Utils';
+import { collectErrorsForMapChecker } from '../utils/MapChecker.Utils';
 import './ReactFlowStyleOverrides.css';
 import { ReactFlowWrapper } from './ReactFlowWrapper';
 import { Stack } from '@fluentui/react';
@@ -116,6 +117,7 @@ export const DataMapperDesigner = ({
   const isMapStateDirty = useSelector((state: RootState) => state.dataMap.isDirty);
   const sourceSchema = useSelector((state: RootState) => state.dataMap.curDataMapOperation.sourceSchema);
   const targetSchema = useSelector((state: RootState) => state.dataMap.curDataMapOperation.targetSchema);
+  const flattenedTargetSchema = useSelector((state: RootState) => state.dataMap.curDataMapOperation.flattenedTargetSchema);
   const currentTargetSchemaNode = useSelector((state: RootState) => state.dataMap.curDataMapOperation.currentTargetSchemaNode);
   const currentConnections = useSelector((state: RootState) => state.dataMap.curDataMapOperation.dataMapConnections);
   const selectedItemKey = useSelector((state: RootState) => state.dataMap.curDataMapOperation.selectedItemKey);
@@ -126,7 +128,7 @@ export const DataMapperDesigner = ({
   const [isCodeViewOpen, setIsCodeViewOpen] = useState(false);
   const [isTestMapPanelOpen, setIsTestMapPanelOpen] = useState(false);
   const [isTargetSchemaPaneExpanded, setIsTargetSchemaPaneExpanded] = useState(false);
-  const [isMapCheckerOpen, { setFalse: closeMapChecker, toggle: toggleMapChecker }] = useBoolean(false);
+  const [isMapCheckerOpen, { setTrue: openMapChecker, setFalse: closeMapChecker, toggle: toggleMapChecker }] = useBoolean(false);
 
   const dataMapDefinition = useMemo<string>(() => {
     if (sourceSchema && targetSchema) {
@@ -167,6 +169,18 @@ export const DataMapperDesigner = ({
   };
 
   const onSaveClick = useCallback(() => {
+    const errors = collectErrorsForMapChecker(currentConnections, flattenedTargetSchema);
+
+    if (errors.length > 0) {
+      dispatch(
+        showNotification({
+          type: NotificationTypes.MapHasErrorsAtSave,
+          msgParam: errors.length,
+          autoHideDurationMs: errorNotificationAutoHideDuration,
+        })
+      );
+    }
+
     generateDataMapXslt(dataMapDefinition)
       .then((xsltStr) => {
         saveStateCall(dataMapDefinition, xsltStr);
@@ -195,7 +209,7 @@ export const DataMapperDesigner = ({
           })
         );
       });
-  }, [dispatch, dataMapDefinition, saveStateCall, sourceSchema, targetSchema]);
+  }, [currentConnections, flattenedTargetSchema, dataMapDefinition, saveStateCall, dispatch, sourceSchema, targetSchema]);
 
   // NOTE: Putting this useEffect here for vis next to onSave
   useEffect(() => {
@@ -302,6 +316,7 @@ export const DataMapperDesigner = ({
                           canvasBlockHeight={getCanvasAreaHeight()}
                           canvasBlockWidth={centerViewWidth}
                           useExpandedFunctionCards={useExpandedFunctionCards}
+                          openMapChecker={openMapChecker}
                         />
                       </ReactFlowProvider>
                     )}