+++ libs/designer/src/lib/core/state/designerOptions/designerOptionsInterfaces.ts
@@ -7,6 +7,7 @@ import type {
   ISearchService,
   IOAuthService,
   IWorkflowService,
+  IHostService,
   IApiManagementService,
   IFunctionService,
   IAppServiceService,
@@ -29,6 +30,7 @@ export interface ServiceOptions {
   loggerService?: ILoggerService;
   oAuthService: IOAuthService;
   workflowService: IWorkflowService;
+  hostService?: IHostService;
   apimService?: IApiManagementService;
   functionService?: IFunctionService;
   appServiceService?: IAppServiceService;